requirements.txt
# FastAPI and Server
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
argon2-cffi==23.1.0

# File Processing
pypdf==5.2.0
python-docx==1.1.0
Pillow==10.3.0

# AI and OCR
google-generativeai==0.3.2
pytesseract==0.3.10

# Validation and Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
loguru==0.7.2
